pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chimera"
version = "0.1.0"
description = "Biologically-inspired OS homeostasis engine"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Terry" }]
dependencies = [
    "psutil>=5.9",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "structlog>=24.1",
    "numpy>=1.26",
]

[project.optional-dependencies]
windows = [
    "pywin32>=306; sys_platform == 'win32'",
    "wmi>=1.5; sys_platform == 'win32'",
]
ui = [
    "pystray>=0.19",
    "Pillow>=10.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.26",
]
llm = [
    "anthropic>=0.39",
    "ollama>=0.2",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.8",
    "hypothesis>=6.98",
]

[project.scripts]
chimera = "chimera.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/chimera"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "windows: tests that require Windows-specific APIs",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["chimera"]
mypy_path = "src"
